refactor: shared document_state
