+++ packages/crdt_lf/flutter_example/lib/todo_list/state.dart
@@ -1,46 +1,18 @@
-import 'dart:async';
-
 import 'package:crdt_lf/crdt_lf.dart';
+import 'package:crdt_lf_flutter_example/shared/document_state.dart';
 import 'package:crdt_lf_flutter_example/shared/network.dart';
-import 'package:flutter/material.dart';
 
-class DocumentState extends ChangeNotifier {
-  DocumentState._(this._document, this._handler, this._network) {
-    _listenToNetworkChanges();
-    _listenToLocalChanges();
-  }
+class TodoDocumentState extends DocumentState {
+  TodoDocumentState._(CRDTDocument document, this._handler, Network network)
+    : super(document, network);
 
-  factory DocumentState.create(PeerId author, {required Network network}) {
+  factory TodoDocumentState.create(PeerId author, {required Network network}) {
     final document = CRDTDocument(peerId: author);
     final handler = CRDTListHandler<String>(document, 'todo-list');
-    return DocumentState._(document, handler, network);
+    return TodoDocumentState._(document, handler, network);
   }
 
-  final CRDTDocument _document;
   final CRDTListHandler<String> _handler;
-  final Network _network;
-
-  StreamSubscription<Change>? _networkChanges;
-  StreamSubscription<Change>? _localChanges;
-
-  void _listenToNetworkChanges() {
-    _networkChanges = _network
-        .stream(_document.peerId)
-        .listen(_applyNetworkChanges);
-  }
-
-  void _listenToLocalChanges() {
-    _localChanges = _document.localChanges.listen(_sendChange);
-  }
-
-  void _applyNetworkChanges(Change change) {
-    _document.applyChange(change);
-    notifyListeners();
-  }
-
-  void _sendChange(Change change) {
-    _network.sendChange(change);
-  }
 
   void addTodo(String todo) {
     _handler.insert(0, todo);
@@ -53,11 +25,4 @@ class DocumentState extends ChangeNotifier {
   }
 
   List<String> get todos => _handler.value;
-
-  @override
-  void dispose() {
-    _networkChanges?.cancel();
-    _localChanges?.cancel();
-    super.dispose();
-  }
 }
